backend/workbench/requirements.txt
asgiref==3.6.0
Django==4.2.1
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==23.2
django-http-method==1.2.1
django-rename-app==0.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
graphviz==0.20.1
Markdown==3.4.3
numpy==1.24.3
Pillow==9.5.0
pydotplus==2.0.2
PyJWT==2.6.0
pyparsing==3.0.9
pytube==12.1.3
pytz==2023.3
PyYAML==6.0
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
tzdata==2023.3
